+++ app/src/androidTest/java/com/fixupxer/CustomRulesUiTest.kt
@@ -11,18 +11,25 @@
 
 package com.fixupxer
 
+import android.content.Context
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
 import androidx.test.core.app.ActivityScenario
+import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.action.ViewActions.closeSoftKeyboard
 import androidx.test.espresso.action.ViewActions.replaceText
 import androidx.test.espresso.action.ViewActions.swipeUp
+import androidx.test.espresso.assertion.ViewAssertions.doesNotExist
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
+import androidx.test.espresso.matcher.ViewMatchers.isNotChecked
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.fixupxer.ui.SettingsActivity
+import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -30,14 +37,33 @@ import org.junit.runner.RunWith
 class CustomRulesUiTest {
     @Test
     fun createRuleFromSettingsWithoutRawJson() {
+        val context = ApplicationProvider.getApplicationContext<Context>()
+        PreferencesManager(context).setCustomRulesEnabled(false)
+
         ActivityScenario.launch(SettingsActivity::class.java).use {
             onView(withId(R.id.buttonCustomRulesHowTo)).check(matches(isDisplayed()))
+            onView(withId(R.id.switchCustomRules)).check(matches(isNotChecked()))
             onView(withId(R.id.buttonCustomRules)).perform(click())
             onView(withText(R.string.custom_rules_title)).check(matches(isDisplayed()))
+            onView(withId(R.id.switchCustomRules)).check(doesNotExist())
+            onView(withId(R.id.buttonAddRule)).check { view, noViewFound ->
+                noViewFound?.let { throw it }
+                val location = IntArray(2)
+                view.getLocationOnScreen(location)
+                val navigationInset = ViewCompat.getRootWindowInsets(view)
+                    ?.getInsets(WindowInsetsCompat.Type.navigationBars())
+                    ?.bottom ?: 0
+                val usableBottom = view.rootView.height - navigationInset
+                assertTrue(
+                    "Add rule button overlaps the navigation bar",
+                    location[1] + view.height <= usableBottom
+                )
+            }
             onView(withId(R.id.buttonAddRule)).perform(click())
             onView(withId(R.id.editName))
                 .perform(replaceText("Instrumentation rule"), closeSoftKeyboard())
-            onView(withId(R.id.mainScrollView)).perform(swipeUp(), swipeUp(), swipeUp())
+            onView(withId(R.id.mainScrollView))
+                .perform(swipeUp(), swipeUp(), swipeUp(), swipeUp(), swipeUp())
             onView(withId(R.id.buttonSave)).perform(click())
             onView(withText("Instrumentation rule")).check(matches(isDisplayed()))
             onView(withId(R.id.buttonClear)).perform(click())